# swagger.yaml
openapi: 3.0.3
info:
  title: Document Processing Wizard API
  description: |
    AI-powered document processing with wizard flow.

    ## Features
    - Multi-format document upload (PDF, DOCX, XLSX)
    - Priority-based prompt processing
    - Multiple AI providers (Ollama, OpenAI, Gemini)
    - Conversation persistence with clarifications
    - Version control and document modification
    - Czech language support

    ## Workflow
    1. **Upload Files** - Upload documents and get session ID
    2. **Submit Prompts** - Define processing tasks with priorities
    3. **Monitor Status** - Track processing progress
    4. **Handle Clarifications** - Answer AI questions if needed
    5. **Review Result** - Get processed document
    6. **Confirm/Modify** - Approve or request changes

  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
...
paths:
...
  /wizard/session/{sessionId}:
    get:
      tags:
        - Status & Info
      summary: Get Session Details
      description: Get complete session information including files, prompts, and results
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/models/sync:
    post:
      tags:
        - Admin
      summary: Sync Models from Ollama
      description: Query Ollama instance and sync available models to database
      operationId: syncModels
      responses:
        '200':
          description: Models synced successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Models synced successfully
                  data:
                    type: object
                    properties:
                      totalModels:
                        type: integer
                        example: 3
                      availableModels:
                        type: integer
                        example: 3
                      enabledModels:
                        type: integer
                        example: 3
                      models:
                        type: array
                        items:
                          $ref: '#/components/schemas/OllamaModel'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/models:
    get:
      tags:
        - Admin
      summary: List All Models
      description: Get all Ollama models from database
      operationId: listModels
      parameters:
        - name: includeUnavailable
          in: query
          description: Include models not currently available in Ollama
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Models retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      models:
                        type: array
                        items:
                          $ref: '#/components/schemas/OllamaModel'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/models/recommended:
    get:
      tags:
        - Admin
      summary: Get Recommended Model
      description: Get the best available model based on priority and availability
      operationId: getRecommendedModel
      responses:
        '200':
          description: Recommended model retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      modelName:
                        type: string
                        example: llama3.1:8b
                      message:
                        type: string
                        example: This model will be used for processing
        '404':
          description: No available models
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No Available Models
                  message:
                    type: string
                    example: No enabled models found. Please sync or enable models.
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/models/pull:
    post:
      tags:
        - Admin
      summary: Pull Model from Ollama
      description: Download a new model from Ollama registry (runs in background)
      operationId: pullModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - modelName
              properties:
                modelName:
                  type: string
                  example: llama3:8b
                  description: Name of the model to pull
      responses:
        '200':
          description: Model pull initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Started pulling model llama3:8b
                  data:
                    type: object
                    properties:
                      modelName:
                        type: string
                        example: llama3:8b
                      status:
                        type: string
                        example: pulling
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/models/{id}:
    patch:
      tags:
        - Admin
      summary: Update Model Configuration
      description: Update model settings (priority, enabled status, etc.)
      operationId: updateModel
      parameters:
        - name: id
          in: path
          required: true
          description: Model UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  example: Llama 3.1 (8B)
                isEnabled:
                  type: boolean
                  example: true
                priority:
                  type: integer
                  example: 10
                  description: Lower number = higher priority
                maxTokens:
                  type: integer
                  example: 4096
                temperature:
                  type: number
                  format: float
                  example: 0.7
      responses:
        '200':
          description: Model updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Model updated successfully
                  data:
                    type: object
                    properties:
                      model:
                        $ref: '#/components/schemas/OllamaModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Admin
      summary: Delete Model from Database
      description: Remove model from database (does not delete from Ollama)
      operationId: deleteModel
      parameters:
        - name: id
          in: path
          required: true
          description: Model UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Model deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Model removed from database
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Session UUID
      schema:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
  schemas:
...
    OllamaModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the model
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: Ollama model name (e.g., llama3.1:8b)
          example: llama3.1:8b
        displayName:
          type: string
          description: User-friendly display name
          example: Llama 3.1 (8B)
        size:
          type: string
          description: Model size in bytes (as string due to BigInt)
          example: "4700000000"
        family:
          type: string
          description: Model family (e.g., llama, mistral)
          example: llama
        parameterSize:
          type: string
          description: Parameter size (e.g., 7B, 8B, 70B)
          example: 8B
        quantization:
          type: string
          description: Quantization level (e.g., Q4_0, Q5_1)
          example: Q4_0
        isAvailable:
          type: boolean
          description: Whether model is currently available in Ollama
          example: true
        isEnabled:
          type: boolean
          description: Whether model is enabled for use
          example: true
        priority:
          type: integer
          description: Model priority (lower = higher priority)
          example: 10
        maxTokens:
          type: integer
          description: Maximum context tokens
          example: 4096
          nullable: true
        temperature:
          type: number
          format: float
          description: Default temperature setting
          example: 0.7
          nullable: true
        usageCount:
          type: integer
          description: Number of times model has been used
          example: 42
        lastUsed:
          type: string
          format: date-time
          description: Timestamp of last usage
          nullable: true
        lastChecked:
          type: string
          format: date-time
          description: Timestamp of last availability check
        createdAt:
          type: string
          format: date-time
          description: Timestamp when model was added to database
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of last update
...
    SessionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            session:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                status:
                  type: string
                createdAt:
                  type: string
                  format: date-time
                expiresAt:
                  type: string
                  format: date-time
                files:
                  type: array
                  items:
                    type: object
                prompts:
                  type: array
                  items:
                    type: object
                results:
                  type: array
                  items:
                    type: object
            conversation:
              type: object
              properties:
                totalMessages:
                  type: integer
                clarificationCount:
                  type: integer
                lastMessageAt:
                  type: string
                  format: date-time
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Validation error
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Session not found
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Internal server error
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (not yet implemented)

# Uncomment when authentication is implemented
# security:
#   - ApiKeyAuth: []